separated nested classes wherever found
