+++ legacy/Modules/DataModule.cs
@@ -1,6 +1,4 @@
-﻿//using System;
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.ComponentModel.DataAnnotations;
@@ -1605,36 +1603,5 @@ namespace Orikivo.Modules
 
             await ReplyAsync(embed: e.Build());
             }
-
-        //stats stuff
-        [Name("Analysis")]
-        [Summary("Provides statistics on a range of objects.")]
-        public class AnalyticsModule
-        {
-            public AnalyticsModule()
-            {
-
-            }
-
-            //[Command("analyze")]
-            public async Task AnalyzeObjectResponseAsync()
-            {
-
-            }
-
-
-
-            /*
-                    [Command("npxlacc")]
-                    [Summary("A rewritten pixel profile tool.")]
-                    public async Task BuildCardAsync()
-                    {
-                        string card = PixelEngine.GetCardAsPath(Context.User, new PixelRenderingOptions(Context.Account));
-                        EmbedBuilder e = new EmbedBuilder();
-                        e.WithImageUrl($"attachment://{Path.GetFileName(card)}");
-                        await Context.Channel.SendFileAsync(card, embed: e.Build());
-                    }
-            */
-        }
     }
 }
